refactoring update flow (broken version)
